+++ src/presentation/handler/tokens_handlers.py
@@ -62,3 +62,9 @@ def delete_token_for_user(user_uuid_id: str, token_name: str):
         },
     )
     return True
+
+
+def set_token_as_default_for_user(user_uuid_id: str, token_name: str):
+    ai_mananger.set_default_token(user_uuid_id=user_uuid_id, token_name=token_name)
+
+    return True