+++ CHANGELOG.md
@@ -69,6 +69,7 @@
 - *(tokens)* Created some routes and handlers for communication
 - *(tokens)* Improved default tokens verification
 - *(routes)* Added the route to the main
+- *(tokens)* Route for set default token
 
 ### 🐛 Bug Fixes